requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
openai==1.51.0
httpx==0.26.0